Enhance UI and error handling in package management
- Improved error handling in MyPackagesFragment to log and display errors when submitting package list.
- Updated PackageAdapter to display creator ID instead of username for better clarity.
- Refined layout styles in fragment_create_package.xml and fragment_edit_package.xml for a more cohesive dark theme.
- Enhanced text input fields with better padding, colors, and hints for improved user experience.
- Added progress bars and adjusted button styles for consistency across the application.
- Updated labels and sections for better organization and clarity in the package creation and editing process.

+++ FrontEnd/app/src/main/java/com/fitapp/appfit/feature/marketplace/model/request/CreatePackageRequest.kt
@@ -11,19 +11,19 @@ data class CreatePackageRequest(
     val description: String?,
 
     @SerializedName("packageType")
-    val packageType: String, // SPORT_PACK, PARAMETER_PACK, ROUTINE_PACK, EXERCISE_PACK, MIXED
+    val packageType: String,
 
     @SerializedName("isFree")
     val isFree: Boolean,
 
     @SerializedName("price")
-    val price: BigDecimal?,
+    val price: Double?,
 
     @SerializedName("currency")
     val currency: String?,
 
     @SerializedName("requiresSubscription")
-    val requiresSubscription: String, // FREE, STANDARD, PREMIUM
+    val requiresSubscription: String,
 
     @SerializedName("thumbnailUrl")
     val thumbnailUrl: String?,
@@ -33,5 +33,4 @@ data class CreatePackageRequest(
 
     @SerializedName("initialItems")
     val initialItems: List<AddPackageItemRequest>
-
 )
